+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next
 
 - Feat: match longest definitions first. See https://github.com/DiscreteTom/bimark/issues/3.
+- Feat: ignore files in `.gitignore`.
 
 ## v0.1.0